+++ crates/zspell/src/system.rs
@@ -1,17 +1,19 @@
-use cfg_if::cfg_if;
-use hashbrown::HashSet;
-use home::home_dir;
-use regex::{escape, Regex};
+use std::collections::HashSet;
 use std::ffi::OsStr;
 use std::fs;
+
 use std::{
     env,
     path::{Component, Path, PathBuf},
 };
-// use sys_locale::get_locale;
 
-use crate::errors::DictError;
+use cfg_if::cfg_if;
+use home::home_dir;
+use regex::{escape, Regex};
+use sys_locale::get_locale;
+
 // use crate::errors::FileError;
+use crate::errors::{DictError, SystemError};
 use crate::{unwrap_or_ret, Dictionary};
 
 pub const PKG_NAME: &str = env!("CARGO_PKG_NAME");
@@ -76,6 +78,18 @@ const CHILD_DIR_NAMES: [&str; 8] = [
     "hunspell",
     "dicts",
 ];
+
+const AFF_EXTENSIONS: [&str; 3] = ["aff", "afx", "affix"];
+const DIC_EXTENSIONS: [&str; 3] = ["dic", "dict", "dictionary"];
+
+/// Get the user's language from their system
+///
+/// Currently this only uses locale and defaults to en-US. Eventually we will
+/// have it adapt to the user's config file in ~/.zspell.
+#[inline]
+pub fn get_preferred_lang() {
+    get_locale().unwrap_or_else(|| String::from("en-US"));
+}
 
 /// Create a list of possible locations to find dictionary files. Expands home;
 /// does not expand windcards
@@ -153,19 +167,10 @@ pub fn find_matching_dirs(parent: &Path, pattern: &str) -> Vec<PathBuf> {
     let matched_iter = dir_items
         // Get Ok() values
         .filter_map(Result::ok)
-        // Looks tricky, but just returns the path if our item is a directory
-        .filter_map(|x| match x.file_type() {
-            Ok(ft) => {
-                if ft.is_dir() {
-                    Some(x.path())
-                } else {
-                    None
-                }
-            }
-            Err(_) => None,
-        })
-        // Get items that match
-        .filter(|x| re.is_match(&x.to_string_lossy()));
+        // Looks tricky, but just filters out anything that isn't a directory
+        .filter(|dir_entry| dir_entry.file_type().map_or(false, |v| v.is_dir()))
+        .map(|dir_entry| dir_entry.path())
+        .filter(|path_buf| re.is_match(&path_buf.to_string_lossy()));
 
     // Create a new item for the parent paths if exists
     for item in matched_iter {
@@ -182,47 +187,55 @@ pub fn find_matching_dirs(parent: &Path, pattern: &str) -> Vec<PathBuf> {
 ///
 /// This takes a mutable vector that will be drained (used as a stack).
 #[inline]
-pub fn expand_dir_wildcards(paths: &mut Vec<PathBuf>) -> HashSet<PathBuf> {
+pub fn expand_dir_wildcards(path_queue: &mut Vec<PathBuf>) -> HashSet<PathBuf> {
     // We will collect only the existing values here
     let mut ret = HashSet::new();
 
     // Work to empty our stack
-    while let Some(top) = paths.pop() {
-        // println!("working path {top:?}");
-        // println!("remaining paths: {:#?}", paths);
-
+    'loop_queue: while let Some(path) = path_queue.pop() {
         // This will hold the "working" parent path
         let mut cur_base = PathBuf::new();
-        let mut is_new = true;
+        let mut comp_iter = path.components();
+        let mut is_first_comp = true;
 
-        for comp in top.components() {
-            // println!("base: {cur_base:?} (new: {is_new}) comp {comp:?}");
+        // The first component will be
+        'loop_comps: while let Some(comp) = comp_iter.next() {
             // If our parent doesn't exist or is not a dir, we're done here
-            if !is_new && (!cur_base.exists() || !cur_base.is_dir()) {
-                // println!("breaking");
-                break;
+            // Don't check this on the first loop when we have an empty buffer
+            if !(is_first_comp || cur_base.exists() && cur_base.is_dir()) {
+                break 'loop_comps;
             }
 
-            if let Component::Normal(value) = comp {
-                // Enter here if this part of the
-                let val_str = value.to_string_lossy();
-                if val_str.contains('*') {
-                    // println!("finding matching");
-                    find_matching_dirs(&cur_base, &val_str);
-                } else {
-                    // println!("adding normal");
-                    // Just add anything else
+            match comp {
+                Component::Normal(value) if value.to_string_lossy().contains('*') => {
+                    // This block handles strings with wildcards
+
+                    // Optimizer should get this reuse
+                    let val_str = value.to_string_lossy();
+                    let remaining_path: PathBuf = comp_iter.collect();
+
+                    // Find directories that match the wildcard
+                    let mut matching_dirs = find_matching_dirs(&cur_base, &val_str);
+
+                    // Append the rest of our buffer to each of them
+                    for matching_path_buf in &mut matching_dirs {
+                        matching_path_buf.push(remaining_path.clone());
+                    }
+
+                    // Save the existing paths to our queue
+                    path_queue.append(&mut matching_dirs);
+                    continue 'loop_queue;
+                }
+                _ => {
+                    // Anything else just gets added on with no fanfare
                     cur_base.push(comp);
                 }
-            } else {
-                // println!("adding other");
-                // Anything else just gets added on with no fanfare
-                cur_base.push(comp);
             }
 
-            is_new = false;
+            is_first_comp = false;
         }
 
+        // Check if our base exists and is valid; if so, add it
         if cur_base.exists() && cur_base.is_dir() {
             ret.insert(cur_base);
         }
@@ -231,9 +244,94 @@ pub fn expand_dir_wildcards(paths: &mut Vec<PathBuf>) -> HashSet<PathBuf> {
     ret
 }
 
-// pub fn find_dict_from_path() {
-//     let locale = get_locale().unwrap_or_else(|| String::from("en-US"));
-// }
+#[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
+struct PathInfo {
+    buf: PathBuf,
+    stem: String,
+    extension: String,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
+pub struct DictPaths {
+    pub dictionary: PathBuf,
+    pub affix: PathBuf,
+}
+
+/// Given a path and a language, find any potential dictionary files
+///
+/// # Errors
+///
+/// If the directory cannot be accessed, return an error. This wraps a
+/// [`std::io::ErrorKind`] error.
+#[inline]
+pub fn find_dicts_from_path<T: AsRef<str>>(
+    path: &Path,
+    lang: T,
+) -> Result<Vec<DictPaths>, SystemError> {
+    let lang_ref = lang.as_ref().to_lowercase();
+
+    // Sometimes we get something that's e.g. en_US and sometimes en-US
+    // Might have duplicates here, no problem since we only use `contains`
+    let loc_bases = [
+        lang_ref.clone(),
+        lang_ref.replace('-', "_"),
+        lang_ref.replace('_', "-"),
+        lang_ref.split(['-', '_']).next().unwrap().to_owned(),
+    ];
+
+    let dir_iter = match fs::read_dir(path) {
+        Ok(v) => v,
+        Err(e) => {
+            return Err(SystemError::IOError {
+                name: path.to_string_lossy().to_string(),
+                e: e.kind(),
+            })
+        }
+    };
+
+    // Collect all paths that are files and are named the location base
+    let possible_paths: Vec<PathInfo> = dir_iter
+        .filter_map(Result::ok)
+        .map(|entry| entry.path())
+        .filter(|path| path.is_file())
+        // Keep the path, file name, and extension together
+        .map(|path| PathInfo {
+            stem: path
+                .file_stem()
+                .unwrap_or_default()
+                .to_string_lossy()
+                .to_lowercase(),
+            extension: path
+                .extension()
+                .unwrap_or_default()
+                .to_string_lossy()
+                .to_lowercase(),
+            buf: path,
+        })
+        // Only get possible matching locale file names
+        .filter(|pinfo| loc_bases.contains(&pinfo.stem))
+        .collect();
+
+    let existing_file_maps: Vec<DictPaths> = possible_paths
+        .iter()
+        // Start with a dictionary file
+        .filter(|pinfo| DIC_EXTENSIONS.contains(&pinfo.extension.as_str()))
+        // Existing same names with an affix file
+        .flat_map(|pinfo| {
+            possible_paths
+                .iter()
+                .filter(|pi| pi.stem == pinfo.stem)
+                .filter(|pi| AFF_EXTENSIONS.contains(&pi.extension.as_str()))
+                // Put matches into a struct
+                .map(|pi| DictPaths {
+                    dictionary: pinfo.buf.clone(),
+                    affix: pi.buf.clone(),
+                })
+        })
+        .collect();
+
+    Ok(existing_file_maps)
+}
 
 /// Take in a path and load the dictionary
 ///
@@ -255,7 +353,7 @@ pub fn create_dict_from_path(basepath: &str) -> Result<Dictionary, DictError> {
         Err(e) => {
             return Err(DictError::FileError {
                 fname: affix_file_path,
-                orig_e: e.kind(),
+                e: e.kind(),
             })
         }
     }
@@ -265,7 +363,7 @@ pub fn create_dict_from_path(basepath: &str) -> Result<Dictionary, DictError> {
         Err(e) => {
             return Err(DictError::FileError {
                 fname: dict_file_path,
-                orig_e: e.kind(),
+                e: e.kind(),
             })
         }
     }
@@ -280,7 +378,8 @@ pub fn create_dict_from_path(basepath: &str) -> Result<Dictionary, DictError> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::fs;
+    use crate::errors;
+    use std::{fs, io};
     use tempfile::tempdir;
 
     #[test]
@@ -315,9 +414,9 @@ mod tests {
         let dir = tempdir().unwrap();
 
         let mut paths = vec![
-            dir.path().join("a/b/c-x-cxd"),
-            dir.path().join("a/b/c-yz-cxd"),
-            dir.path().join("a/b/c-.abc-cxd"),
+            dir.path().join("a").join("b").join("c-x-cxd"),
+            dir.path().join("a").join("b").join("c-yz-cxd"),
+            dir.path().join("a").join("b").join("c-.abc-cxd"),
         ];
         paths.sort();
 
@@ -325,37 +424,90 @@ mod tests {
             fs::create_dir_all(path).unwrap();
         }
 
-        let mut ret = find_matching_dirs(&dir.path().join("a/b"), "c-*-c?d");
+        let mut ret = find_matching_dirs(&dir.path().join("a").join("b"), "c-*-c?d");
         ret.sort();
 
         assert_eq!(paths, ret);
     }
-    // Test for expand_dir_wildcards
-    // #[test]
-    // fn test_matching_dirs() {
-    //     // Create a temporary directory with contents
-    //     // Ensure the function locates them using wildcards
-    //     let dir = tempdir().unwrap();
 
-    //     let paths = vec![
-    //         dir.path().join("aaa/bbb-x/ccc"),
-    //         dir.path().join("aaa/bbb-y/ccc"),
-    //         dir.path().join("ddd"),
-    //     ];
+    #[test]
+    fn test_expand_dir_wildcards() {
+        let dir = tempdir().unwrap();
 
-    //     for path in &paths {
-    //         fs::create_dir_all(path).unwrap();
-    //     }
+        let paths = vec![
+            dir.path().join("aaa").join("bbb-x").join("ccc"),
+            dir.path().join("aaa").join("bbb-y").join("ccc"),
+            dir.path().join("ddd"),
+        ];
 
-    //     let mut input = vec![
-    //         PathBuf::from("aaa/bbb*/ccc"),
-    //         PathBuf::from("ddd"),
-    //     ];
+        for path in &paths {
+            fs::create_dir_all(path).unwrap();
+        }
 
-    //     let mut expanded = Vec::from_iter(expand_dir_wildcards(&mut input));
-    //     expanded.sort_unstable();
+        let mut input = vec![
+            dir.path().join("aaa").join("bbb*").join("ccc"),
+            dir.path().join("ddd"),
+        ];
 
-    //     assert_eq!(paths, expanded);
+        let mut expanded = Vec::from_iter(expand_dir_wildcards(&mut input));
+        expanded.sort_unstable();
 
-    // }
+        assert_eq!(paths, expanded);
+    }
+
+    #[test]
+    fn test_find_dict_from_path() {
+        let dir = tempdir().unwrap();
+
+        let fnames = vec![
+            dir.path().join("test_found.dic"),
+            dir.path().join("test_found.aff"),
+            dir.path().join("test_found.afx"),
+            dir.path().join("test.dict"),
+            dir.path().join("test.affix"),
+            dir.path().join("notfound.dic"),
+            dir.path().join("notfound.aff"),
+            dir.path().join("test"),
+        ];
+
+        let mut expected = vec![
+            DictPaths {
+                dictionary: fnames[0].clone(),
+                affix: fnames[1].clone(),
+            },
+            DictPaths {
+                dictionary: fnames[0].clone(),
+                affix: fnames[2].clone(),
+            },
+            DictPaths {
+                dictionary: fnames[3].clone(),
+                affix: fnames[4].clone(),
+            },
+        ];
+        expected.sort();
+
+        for fname in fnames {
+            fs::File::create(fname).unwrap();
+        }
+        fs::read_dir(dir.path()).unwrap();
+
+        let mut res = find_dicts_from_path(dir.path(), "test_found").unwrap();
+        res.sort();
+
+        assert_eq!(res, expected);
+    }
+
+    #[test]
+    fn test_find_dict_from_path_err() {
+        let fakepath = tempdir().unwrap().path().join("fake");
+        let res = find_dicts_from_path(&fakepath, "test_found");
+
+        assert_eq!(
+            Err(errors::SystemError::IOError {
+                name: fakepath.to_string_lossy().to_string(),
+                e: io::ErrorKind::NotFound
+            }),
+            res
+        )
+    }
 }